+++ product_code/certificate_service/impl/src/main/scala/de/upb/cs/uc4/certificate/impl/CertificateApplication.scala
@@ -9,9 +9,10 @@ import com.lightbend.lagom.scaladsl.persistence.slick.SlickPersistenceComponents
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
 import com.lightbend.lagom.scaladsl.server.{ LagomApplicationContext, LagomServer }
 import com.softwaremill.macwire.wire
+import de.upb.cs.uc4.authentication.model.JsonUsername
 import de.upb.cs.uc4.certificate.api.CertificateService
 import de.upb.cs.uc4.certificate.impl.actor.{ CertificateBehaviour, CertificateState }
-import de.upb.cs.uc4.certificate.impl.commands.RegisterUser
+import de.upb.cs.uc4.certificate.impl.commands.{ DeleteCertificateUser, RegisterUser }
 import de.upb.cs.uc4.certificate.impl.readside.CertificateEventProcessor
 import de.upb.cs.uc4.hyperledger.HyperledgerAdminParts
 import de.upb.cs.uc4.hyperledger.HyperledgerUtils.ExceptionUtils
@@ -82,6 +83,25 @@ abstract class CertificateApplication(context: LagomApplicationContext)
           log.error("CertificateService received invalid topic message: {}", throwable.toString)
           Future.successful(Done)
       }).mapAsync(8)(done => done)
+    )
+
+  userService
+    .userDeletionTopic()
+    .subscribe
+    .atLeastOnce(
+      Flow.fromFunction[EncryptionContainer, Future[Done]] { container =>
+        try {
+          val jsonUsername = kafkaEncryptionUtility.decrypt[JsonUsername](container)
+          clusterSharding.entityRefFor(CertificateState.typeKey, jsonUsername.username)
+            .ask[Confirmation](replyTo => DeleteCertificateUser(jsonUsername.username, replyTo)).map(_ => Done)
+        }
+        catch {
+          case throwable: Throwable =>
+            log.error("AuthenticationService received invalid topic message: {}", throwable.toString)
+            Future.successful(Done)
+        }
+      }
+        .mapAsync(8)(done => done)
     )
 
   private def registerUser(username: String, enrollmentId: String): Future[Done] = {